InputWidget and subclasses: Remember original value when creating the widget
The first time that the value is set (probably while constructing the
widget), remember it. It can be later used to check whether the value
of the input has been changed since it was created.

We also set the appropriate DOM node properties, so that we are
compatible with code ignorant of OOUI (meant to work with plain HTML
forms).

This previously worked some widgets when infused from PHP (e.g.
TextInputWidget, CheckboxInputWidget), but only by accident, because
we reuse the DOM nodes from HTML. It is now actually supported.

* RadioSelectInputWidget: Use <input class="oo-ui-element-hidden">
  instead of <input type="hidden">, because hidden inputs do not have
  separate 'value' and 'defaultValue' properties.
* CheckboxMultiselectInputWidget: Introduce a similar solution to the
  code previously used in DropdownInputWidget.

Bug: T181691
Change-Id: I30828691af90db6434244229a84e451c85652d74