MenuTagMultiselectWidget: Highlight first item when filtering
When the menu opens, and when we change filtering, and whenever else
the item list in the menu changes, the first item should be highlighted
so the arrow movement can start from the top visually, and so that the
user can press "enter" to add the first given result.

This is similar behavior to LookupElement. Similarly, in LookupElement
the selection happens outside the MenuSelectWidget, rather than inside
its internal operations, since MenuSelectWidget's behavior with arrows
and item visibility iteracts with internal events that prevent properly
highlighting the actual first item in the list.

Bug: T187398
Change-Id: Iff0e9f354b180db23eed5a0791ae8aff069f832d