TagMultiselectWidget: Fix arrow movement in inline input
Inline input grabbed the arrow movement incorrectly, which meant
that we couldn't move through the string properly. With this fix,
the movement is working up to the input edge, which is then transfered
to the tags.

There is a behavior change in this commit after consultation with
Volker: In inline inputs, the "edges" -- either of the
first tag (edge left in LTR) or the end of the string in the input
(edge right in LTR) are defined as boundaries, and the movement
stops rather than wraps back to the previous/later item.

Bug: T189063
Change-Id: I0592269b1416fe230a1cce48cbf243f394d646af