DropdownInput-/RadioSelectInputWidget: Fix support for 'tabIndex'
Support for 'tabIndex' in DropdownInputWidget/RadioSelectInputWidget
has never worked; TabIndexedElement was mixed in, but the 'tabindex'
attribute would be set on the hidden non-interactive input element,
instead of the visible interactive "handle".

Recently, change 06d843c78ce2add47392f40de975435e201c0013 removed the
useless 'tabindex' attribute from them, but we should instead add it
to the actual interactive elements.

Note that this is a bit of a hack still… If anyone calls setTabIndex()
on this.dropdownWidget, weird things may happen, since only one OOUI
widget should be managing this attribute on a particular DOM element.
That does not happen at least in our widgets, though.

Bug: T189610
Change-Id: I486f2c2ff5d47b6abae3fba76961dc1984ff1228