WikimediaUI theme: Don't over-ride selected MenuToolGroup tools' icon
Remove checkmark overriding the normal tool icon in MenuToolGroup when
the tool is selected. This was a very weird design. Selected tool is
already adequately highlighted by the blue color and by changing the
label of the menu.

Also, correct a `background-color` rule that only applied to a single
specific tool in the demo… I think it's supposed to apply everywhere.

Change-Id: I759f9fa7c34b13610be894e16246570ef0985cee