[BREAKING CHANGE] TagItemWidget: Replace 'disabled' items with 'fixed'
For consistency of behavior and separation of UX concerns, don't
allow for individual tag items in a TagMultiselectWidget to be
individually disabled. Instead, allow defining 'fixed' values
that cannot be removed, which is the original intent of the
request to allow for proper behavior with disabled individual
items.

'Disabled' is a state of the entire widget, while 'fixed' items can
be individually set.

Bug: T193571
Change-Id: I312e4e2e4a8f7ec0a06148652a1411bfddb3f8f3