Improve test output in case of failures
Previously we were JSON-encoding the actual and expected data before
passing it to QUnit.equiv(), as a workaround for some weird Karma
issue (597ae95f26bf0a32dd52164fc028f1daadaab270) and as a workaround
for limited nesting depth (fd530a65ca82fbeaf17d1e925b3f086c7983a9d3).

Unfortunately this causes the result to be double-escaped (or even
triple-escaped, in case of the nested HTML code).

However, the Karma issue seems to have been fixed, and I just found
out that the nesting depth limit can be overridden, so we can get rid
of this! (And then we can get rid of some more code, because the
object keys are sorted by default by QUnit.)

Change-Id: I8ca356b6cc6bb13641d18f29623b277e972927c6