MenuSelectWidget: Move 'highlight first item' to end of operation
It is easier to reason about this code if we move this outside of
the loop that marks items visible. Finish the operation on all items
and only afterwards take the first visible item and highlight it.

We make sure that the updating of the item visibility (and the
selection of the top item) is only done if there was an actual
change to the filtering query. Otherwise, the user can and should
be able to select and keep as selected their own item. This fixes
navigation of the menu with up/down arrow keys.

Also, add a 'highlightOnFilter' example in the
MenuTagMultiselectWidget in the demos to show this selection
behavior even when it is inside another widget.

Bug: T190156
Change-Id: I3f6c916de28f0bf70d9e8a9b7d6f5735131954eb