Allow JS/PHP comparison tests for FieldLayout 'help' config option
Even though the PHP and JS implementations are different, we can still
compare some aspects - e.g. the placement of the button in the DOM.
We can also compare it completely when 'helpInline' is set to true.

Remove the exception for FieldLayout 'help', and instead ignore the
contents and other CSS classes of the "oo-ui-fieldLayout-help" element
when comparing.

Also, remove the exception for FieldsetLayout 'help', which has no
implementation in PHP anyway, and so no tests are generated for it.

Change a condition in PHP FieldLayout so that no help button/label are
generated when the parameter is given as empty string. This makes it
consistent with JS FieldLayout and fixes a failure in the new tests.

Change-Id: I7ac4db39d53120dfe1910387b756e65207216a0a