FieldLayout: Reduce clutter in initialization function
The logic of adding the $help element and the necessary attributes
was taking up a lot of space in the init function.

Moving it to a separate function. This also makes the code cleaner
for adding the aria-describedby attribute for both cases: with
helpInline and without.

Follows-up on I90a0a78750c85f070458c385d3b37634b010d210.

Change-Id: I6e4e03f599cf412fc0e3e9a3f9c18ae5d2484e29