Apex theme: Restore space between inline FieldLayout field and label
In change d2b61bee2ea12a36849fb5d9c8757240d5352e9f, this 'padding' was
changed to 'margin', effectively removing the spacing entirely,
because this element is using 'display: table-cell', which ignores all
margins.

Change it back to 'padding', and since the intent of that change was
to reduce vertical spacing, remove 'padding-top' and 'padding-bottom',
which leaves us with only 'padding-left'.

Change-Id: Iaf5ea56c5a3c38317a6bb92f0d87cdada2c19031