MenuTagMultiselectWidget: Cascade disable state to menu
The menu needs to be disabled/enabled when the widget is, as well.
This is most visible when the widget in general starts as disabled
(from PHP; infused) and so the menu is disabled as well, but then
running setDisabled( false ) only enables the input and tag area
and does not enable the menu and items.

Added a disabled widget demo as well.

Bug: T200226
Change-Id: I3962e0289cbeb6d63355511c111d002a64363f55