Fix LookupElement empty search result menu
There appears to be an issue with how the LookupElement clears its
result menu box, this issue was found as part of the AdvancedSearch
extension. Whenever users use backspace to clear all search results or
click on a search result, the result menu box clears its content but it
does not hide itself. This leaves behind an empty box below/above the
input field.

This patch resolves the issue by specifically hiding the menu box in
these cases.

Bug: T189953
Change-Id: I6fa5173220fc5423301ee03ccdeecc3600bee211