NumberInputWidget: Rethink 'step' semantics
Prior to 31e24f02508b, the 'step' config option was only used by the
plus/minus buttons to adjust the value. 'allowInteger' could be used
to validate that the input value is a multiple of 1.

After that change, where we used it for the 'step' HTML attribute, it
suddenly also started to be used to validate that the input value is a
multiple of the step, making 'allowInteger' superfluous.

After this change:
* 'step' is still used to validate that the input value is a multiple
  of the step, but now we document this behavior. It now defaults to
  null (no validation) rather than 1.
* New option 'buttonStep' is added for the plus/minus button
  adjustment, defaulting to 'step', or 1 if not given.
* 'allowInteger' is soft deprecated (like its old alias 'isInteger'),
  now it implies `step: 1`.
* setStep() now takes a third parameter to set the validity step, and
  getStep() returns it as the third element of the array.

Bug: T203766
Change-Id: I8981ee1185b9ebf658e26e6ff4e3433cf632b6cc