Remove unnecessary empty-theme.less
This file existed "so that custom themes don't need to define tons of
empty mixins", but in practice both of our themes did exactly that and
redefined all the empty mixins (except one, which was clearly forgotten).

I think it is better to be explicit about it. A missing mixin should
cause the styles to not build, rather than silently keep going.

Change-Id: I87af3d84c7e1391de48109b03c836e63955260e9