Centralize the definition of which classes belong in which module
Previously, we defined which JS files are used to build each module
(core, widgets, toolbars, windows) in build/modules.yaml, but we
defined which Less files are used in four separate .less files in
src/styles/. Move that to build/modules.yaml too.

Output is identical except for a few extra empty lines.

Change-Id: I87a2197878fc40d5786c03e4d3519314807fb338