Fix some typos in unimportant places
I wrote a script to find every word that is only used once throughout
the project (JavaScript and PHP files) and which is not an English
word, JavaScript method or PHP function. I reviewed the ~300 results
by hand and found 15 typos. None of them affect how the code works.

I'm not going to share the script publicly because that would curse me
into maintaining it forever. The mind boggles why a proper tool like
that does not exist.

Change-Id: I491ad26faa33d7b31662c76b1770c32a5cc39e53