Allow setting the label to "0" in PHP
In PHP, "0" is falsy (string consisting of the character zero), unlike
in JavaScript, so we were incorrectly treating this value as if no
label was set. We need to check if the label is an empty string rather
than checking its truthiness.

This is a common mistake, so add this to the autogenerated test cases.

Fix a similar bug about FieldLayout 'help', found by the new tests.

Bug: T206925
Change-Id: I354fa4559952ff5e5b9417768b66b2014322c0a9