LabelElement: Allow invisible accessibility labels
When given `label: …, invisibleLabel: true`, widgets using LabelElement
will look as if no label was given, but the label will be present in
generated DOM/HTML, where it is accessible to screen reader software.

This is written with ButtonWidget in mind (and that's the only widget
I manually tested), but it should work for other widgets as well.

IconWidget and IndicatorWidget now also mix in LabelElement, with
`invisibleLabel: true` always set (so they can't have visible labels).

Custom solution for this in FieldLayout/FieldsetLayout is replaced.

Also updating demos to provide labels where useful.

Bug: T94749
Change-Id: Ibe84ec7e89cbc74c55e3682c11a5b281f2671d78