TagMultiSelectWidget: Add a 'limit' configuration option
Add the ability to set a limit on the number of items that can be
selected through TagMultiselectWidget and its children (Menu* and
Popup*).

When the limit is reached, the widget will not accept any more
items and disable its input. If a placeholder is set, it will
be hidden and reappear (and the input re-enabled) only if items
are taken off.

Bug: T208627
Change-Id: I6df5c02039f336efb5117c32d6f473cdcd70e5f1