TagMultiselectWidget: Make widget invalid if there's text in input
Users accidentally leave text in the input, making it confusing
about whether the text inside the input is included in the final
values.

The widget does not consider that text as part of its values, and
does not run any validation on it, so we should make it clear to
users that the leftover text is not valid and either delete it
or add it as a tag.

This commit makes the widget invalid if invalid text was left
in the input, and, since there was no 'invalid' CSS set for the
widget, it sets that as well, with the already existing styles
from the tag widget, and while emphasizing that the problem is
with the leftover text inside the input.

Also add similar treatment of 'outlined' TagMultiselectWidgets
and vars in Apex theme as WikimediaUI theme has been
featuring for a while.

If valid text was left in the widget, this is added as a new
tag and the widget is valid. If the text was invalid but
allowDisplayInvalidTags is true, an invalid tag is added.

Bug: T208507
Bug: T188886
Change-Id: I9d024cd8c2e2b730ab8131a6d5711d69de399cd3