FloatableElement: Remove check for 'needsCustomPosition'
Turns out that code in PopupWidget relied on the bug I fixed in commit
aa43f6ddf97610104c3b2c831495926e776eab45 (by setting 'verticalPosition'
and 'horizontalPosition' to the correct values in 'computePosition()',
which is now not being called when they are set to the initial values).

That code looks difficult to disentangle. Since everything else was
working perfectly correctly before that fix, when this condition was
always true, let us just remove it.

Bug: T208595
Change-Id: I407b0d6fe7b81465054b640d4b5ac4bf352a9901