MenuTagMultiselectWidget: Allow adding arbitrary values
The filtering option meant that when items in the menu were
shown, the menu automatically highlights the top item, so that
it can be inserted. However, when we have the case of a menu
with 'allowArbitrary' then we can't assume the intent is to
add the first value found -- it might be that the user wants
to add an actual arbitrary value. Hitting "enter" adds the
highlighted item, so we need to make sure the right item
is highlighted -- or that no items are.

For this, we prevent the menu from highlighting the first item
if the widget is set to allowArbitrary. We're also making sure
that if there are no visible items after filtering, the highlight
is nulled.

Bonus: Actually add a menu with items to the MenuTagMultiselectWidget
allowArbitrary demo.

Bug: T209717
Change-Id: I7231e8fec6ab3449031b121a188d2e932e0e8268