WindowManager: Prevent iOS Safari from scrolling the page behind the dialog (try#2)
It seems that `overflow: hidden;` is not enough to prevent scrolling.
Adding `height: 100%;` fixes the issue. Also, both properties have to
be set on both <html> and <body>.

This is similar to the approach used in MobileFrontend's overlays.

In doing so, we reset scroll position of the viewport to the top.
Since this is unexpected, and can have confusing effects on sticky
toolbars etc., we can't allow the user to see it happening; so we
don't do the thing for non-fullscreen dialogs.

Bug: T212159
Change-Id: Ic1b449df3184d2430f0f29406c1a8e987084770a