Use `-webkit-overflow-scrolling: touch` for scrollable things
This enables fancy momentum scrolling on iOS devices, most noticeably
in dialogs.

Curious note: when testing (on iPhone SE running iOS 12.1), I found
that `-webkit-overflow-scrolling` is an inherited property (setting it
on a node also enables momentum scrolling in descendant scrollable
nodes, and the value is reported in computed styles for descendants).
But various websites exist that claim it is *not* inherited (e.g.
<https://developer.mozilla.org/en-US/docs/Web/CSS/-webkit-overflow-scrolling>).
Perhaps they are just wrong, but maybe this was true on earlier iOS
versions. Just in case, I'm adding the property everywhere, even if it
would be redundant due to inheritance.

Change-Id: I3f0eb0dff572b26cb4e51941ba970c523e26454e