Make JS/PHP comparison tests async
Group the tests cases into multiple smallers modules and tests, and
add a assert.async() call every 20 tests, so that the next test will
be called asynchronously.

This allows individual modules to be executed in the standalone test,
allows browser to remain interactive while the standalone tests are
running, and I suspect it might also avoid timeouts in Karma tests.

On the other side, it might make the Karma tests slower, but that will
be worth it if it lets them finish consistently without timing out.

Change-Id: I00ea7380fa1571455e95fb1284887b8d0903c889