Deprecation warnings for this.$
OO.ui.Element's this.$ is an alias for the global $ function that has
been deprecated since c38a54272fd7df5ef7aeea4d131d1716ac986a9d (2015),
but it never emitted deprecation warnings. Let's remedy that.

For historical context: OOUI windows used to be an `<iframe>`
containing their contents, and this.$ was a jQuery function that was
"attached" to the document inside that iframe - querying or creating
elements belonging to it. This was known as "window isolation"
and was removed in 0ace0324a150f31713078f4fd91b0f756ecec2d6.

We should remove this.$ rather than leave those harmless 10 characters
around forever because MediaWiki's MobileFrontend also has a this.$
function, that is *not* an alias for $, which is confusing.

Change-Id: I622a0d92b26eafe0816326246825330c7c70aec0