MenuTagMultiselectWidget: 'hideOnChoose' should be set to false
When we have a menu, the menu should never go away unless focus
is out of the widget. Since the widget allows for multiple
choices, the menu should appear even after a choice was already
made.

Also:
- Add this.$group unconditionally to the $autoCloseIgnore
  so that the menu is persistent even if tags are clicked, which
  is its point.
- Do not close the menu when the input is blurred. We want to
  instead have the autoCloseIgnore decide when the toggle the
  menu off.

Note: The entire behavior of the toggling of the menu is
inconsistent, and will be fixed in an upcoming patch in
this chain.

Bug: T163135
Change-Id: Ie5d47dd13ef291c524f69f2c1892c2d05a394f8c