SelectWidget: Allow multiselect mode, add to MenuTagMultiselectWidget
Allow for a multiselect mode in SelectWidget and MenuSelectWidget
and implement it into MenuTagMultiselectWidget for interface and
UX expectation standardization.

This should resolve bugs with UX expectations of the opened menu
whereby the menu displays multiple selections corresponding to the
existing tags in the MenuMultiselectWidget, as well as bugs about
expecting 'selected' single item from a multiselect widget by
actually receiving an array of selections.

Fix up documentKeyDown event to consider multiselect. Also,
there's no real reason for it to consider a selected item,
it should always just find the highlighted item, since it
will then either use it to traverse, or to select/choose.

Bug: T216780
Change-Id: I0946d11de41e2c098893e727b65996f8c66b381d