MenuTagMultiselectWidget: Fix highlight and scrolling to item behavior
With the underlying correction to make the menu persistent, and the
ability to select and unselect menu items which are then connected
to the 'tag' items in the MenuTagMultiselectWidget, we want to make
sure that clicking on a tag item also highlights and then scrolls
to the desired item.

Additional: Fixed a conceptual (and behavioral problem, now that
the menu allows for multiple selections) with assuming that selected
items are needed for keyboard movement. All cases were examined
for SelectWidget#onDocumentKeyDown and in all cases, trusting a
selected item (single or multiple) made no sense as opposed to
trusting either the highlighted item, doing nothing, or using
the first item.

Bug: T209716
Bug: T216071
Change-Id: Iab6e70c590b36117450747762f8fe7000dd9a108