MenuTagMSW: Use 'highlightOnFilter' flag in MenuSelectWidget
Reverts Iff0e9f354b180 which added calls to initializeMenuSelection.

This uses the existing 'highlightOnFilter' functionality, which ensures
the highlight is always done after the menu is filtered.

That functionality is also fixed to ensure that it is only called
if the last highlighted item is no longer visible (or there was none).

Bug: T216780
Change-Id: I929657e59d10571506903c846a6ef3dc2c301db9