Clarify some types in documentation
Technically 'string[]|HtmlSnippet[]' is stricter than what the option really
accepts: it specifies 'an array of strings' or 'an array of HtmlSnippets',
but it actually will accept 'an array of strings or HtmlSnippets'. It would
be more correct to use something like '(string|HtmlSnippet)[]', but we can't
use that syntax.

testsuitegenerator can now deduce the expected values for the changed config
options, so prevent it from generating a bazillion new redundant test cases.

Change-Id: I84a38c91994c3b72099dea3eb6dd288323a1ecb4