Use rounded <rect> elements to optimize some SVG icons
It's not possible to use <rect> elements if there is a hole in the
rectangle. But these don't have holes.

This saves a few bytes both in uncompressed as well as gzipped form.
About -70 bytes uncompressed for all icons in this patch. But that's
not so much the point. I hope this also improves the readability of
the code. Now, it contains a rounded <rect> element for icons that
literally show a rounded rectangle.

Change-Id: Ie05e62f8ce35e408aa1601f85a82fdcec608565c