MenuTagMultiselectWidget: highlightOnFilter only if not `allowArbitrary`
In cases where a MenuTagMSW allows arbitrary values (that are not in
the menu) we don't want to automatically pre-highlight the first
menu item, because we're not *always* filtering. We might want to
add a word in as-is even if it incidentally also filters a few items
in the menu.

highlightOnFilter should work for menu filtering overall, though, so
we will keep its default behavior as "true" for the cases where
allowArbitrary is false, and allow the user to highlight their own
choices (if at all) when arbitrary values are allowed.

Bug: T216071
Change-Id: Ibc4eae40c3d0518a9614b273ef0bc83898a5674a