Recreate settings.svg icon with shorter syntax
I tried to optimize this file manually, just as I did with all the
others. I gave up and ended recreating the image in Inkscape. It
might not be 100% mathematical identical because of this. But I can
not spot a difference even when heavily zoomed in. Please have a look
and tell me if you think this is worth it.

Change-Id: I9c80257a41c89ab21df8d9bcd936fc82c92585f1