Make use of the auto-closing feature in SVG <path>s
The optimization strategy is more or less the same for all these: The
"z" command closes a path. To do this, it is drawing the last missing
stroke as a straight line.

The .svg files I'm touching in this patch all have a path that ends
exactly at the starting point. The "z" command, even if present, was not
doing anything before. It was drawing a line with length "0 0".

I some cases, I could just remove the last command in the path syntax.
But often it was more efficient (= more bytes saved) to change the
starting point of the path.

This is all done manually in PHPStorm.

Change-Id: I40bad4d7525dcfb1d5e17a4f1d9f57ceef30c1af