Manually recreate settings.svg icon
This version keeps the standard viewBox, but makes use of a series
of transformations. Note it's not the only icon in this library that
uses transformations.

Before:           551 raw bytes, 336 bytes gzipped.
After this patch: 540 raw bytes, 302 bytes gzipped.

This is not as good as in patch set 1, where a non-standard viewBox
was used.

Change-Id: Ia47450a2e5afcd2020c3125f54e26022d018261f