Remove full width for CheckboxMultioptionWidget
CheckboxMultioptionWidget is full width, meaning it can be accidentally
selected by clicking on whitespace. On the other hand, CheckboxInputWidget
is not full-width, meaning that the user has to click on the label to
select it. Fix CheckboxMultioptionWidget and give users a consistent
experience.

Follow-up to Ib8661efd035, which does something similar for radio
widgets.

Change-Id: I2a5993e7354b3637a413bb2d622d5691a654aeef