WikimediaUI theme: Use `px` instead of `em`s
Rely on `px` based unit values for everything except
- `font-size` and `line-height`,
- box limiting properties like `width`/`max-width`, `height`/`max-height` &
- inner `padding`s when icons/indicators are involved.

With this, user setting's font size overrides – one of the most used
accessibility features – are still possible and the basic scalability
remains while simplifying coding of interface modifications
in OOUI based products.
It aims for the best-possible compromise between user, design and
developer needs.

This patch also provides `height` properties on all input and other widget
combinations currently part of the library, as relying on `line-height` is
not sufficient in older browsers like IE <= 11 for inputs.

And it fixes a misaligment on TextInputWidget labels in comparison to the
input value itself.

Note, that toolbar and tool calculations remain unchanged as it would
take a much bigger change to amend scaling there. Keeping it to
widgets and layouts for simplicity for now.
Also note, that current calculations in OOUI based products should be
not affected as we are not changing the position or calculated value of any
widget aside of above TextInputWidget labels, which are rarely used and
not overriden to our knowledge in production.

Bug: T220671
Change-Id: Ie7e6a19009168d787130f9ee137d8ceb40cad3ff