demos: Expand max-width, from mostly-arbitrary 62.5em to 68.5715em
The current 50em-inspired width is (far) wider than the great many
of our mobile users will ever see, and is a bit narrower than they
will see on desktop, in general. Let's expand a little for sanity.

Also tweak the width of icons so that they display in five columns,
not four; the text over-flow now affects two or three more icons, a
minor issue compared to the denser display benefit.

Change-Id: I7b6eaa489a7fcbc7c5edd1c680c16b21041ed76b