MessageWidget: Amend icon position and `margin` handling
Message icons need to be put on first line in regards to multiline
messages where signaling icon would start too far down.
This was the case with the FieldLayout messages before, we're
re-establishing the preferred treatment.
Also generalizing and simplifying `margin` handling to allow multiple
sequential messages to appear correctly.

Change-Id: I64d7e939576cda85b7903e7a1afcd01b3b8eb634