demos: Use the new workaround for links to anchors with fixed header everywhere
Commit 11ba7212049f6e6b4a044b69098b924b3d88a736 introduced a new
workaround for links to anchors with fixed header (the actual anchors
are positioned with CSS to be 10em above the visual position), and
used it for links to FieldsetLayouts in the demo.

Previously, we already had a different workaround for links to
FieldLayouts (using JS code to fix scroll position whenever the URL
fragment-part changes), which seems to have gone unnoticed.

I like the new workaround more. Let's replace the old one with it.

Change-Id: Ib1b92deafef9173bcec6d0edb8d455095764bbc3