demos: Make forced scrolling rule compatible with disabling scroll for dialogs
The `overflow-y: scroll` on `html` allowed the page to be scrollable
while a OOUI dialog is open, and opening a dialog is supposed to
prevent scrolling by setting `overflow: hidden` on `body`. Move the
rule to `body` so that the override works as expected.

Follow-up to 2aa75094e3e8c5cf6b9a7ece32436fabe46baf14.

Change-Id: I40e5bf0555e3922677fadf2bc9ad173783947ebe