build: Enforce LESS strict units
We're planning to change many variables defined in `em` to be defined
in `px` instead (T220671). Enable strict units to avoid some mistakes.

Currently, trying to add e.g. `1em + 20px` disregards the incompatible
units and returns `21em`. It's difficult to notice the problem when
the numbers are variables, defined in a completely different file.

It turns out that all the variables have actually been defined in
`1/em` all along (reciprocal ems), which is not an unit in CSS, and
only rendered as `em` due to the loose units. This caused errors like
"Multiple units in dimension. Correct the units or use the unit
function. Bad unit: /em" after enabling strict units. Correct that.

Change-Id: I3baf15fd8e0435fc7a05ba6ff6fef58f7aa55563