Remove taint-check escapes_html* annotations
The methods themselves do not escape $config. The annotations were added
because an older version of taint-check thought that these methods would
cause XSS's. However, taint-check is now smart enough to understand
what's going on. Furthermore, if we keep the annotations the current
version of taint-check thinks that we're actually double-escaping the
content. Hence, remove them altogether.

Bug: T230713
Change-Id: Ic96992363cd4641ea5596ad4fa88022063905323