WikimediaUI theme: Fix ProcessDialog head and foot states & button border
ProcessDialog's LESS was overly complex to fulfill all the states
of dialog head (pending) and button flags and states.
This simplifies things and fixes several glitches.
Only side-effect currently is adding a 1px outline to the bottom of
ProcessDialogs when there's no foot content, but it melts into
the dialog border and box-shadow and seems neglible in comparison to
other benefits.

Bug: T234498
Change-Id: Iffb8b45d12d386cfd17205c59efd244bc3efdfe8