WikimediaUI theme: Align text input/dropdown/button paddings with DSG
While writing Components section for Design Style Guide, a misalignment
in text inputs was uncovered, where the icon carrying version featured
too much whitespace to the value or the placeholder.
All paddings should actually be 30/34/38px equivalents for frameless/
text input/framed button like widgets.
This is the first step, in a second we bump the 37 of menus and tools up
to 38.

Change-Id: Ic1171ae4ed57272338551a83a35b50570d824886