TagMultiselectWidget: Remove unnecessary validity check
In 63efa6f414, an extra check was added to onChangeTags, causing the
widget to be invalid if there is any text in the input. However, this
only needed to be done in onInputBlur to solve the bug.

Performing this check in onChangeTags is inconvenient, because it
means that if a new tag is added without clearing the input first,
the widget will be set to invalid.

Bug: T243077
Change-Id: Ifc497f6929cfd7b37393a5269258bbc65e564578