TagMultiselectWidget: Add tags before clearing the input
Before this, addTagFromInput checks the validity of the input, and
if the input value is valid, it clears the input then calls addTag.
Problems with this:
* addTag already validates the input value, so validation is
  needlessly performed twice
* Some downstream widgets fail validation after the input has been
  cleared (T238318)
* The input should probably not be cleared if the tag was not added,
  even if the input value was valid

After this, addTagFromInput does not perform its own validation
checks, and only clears the input after the tag was successfully
added.

Related to this, ensure MenuTagMultiselectWidget#onMenuChoose adds
tags before clearing the input.

Bug: T243075
Change-Id: I777ff7f193177555e51ca4cff0e6878b48bec250