Improve 'ooui-combobox-button-label' message
This is a screenreader-only message that is supposed to describe the
button in a combobox that opens the dropdown with preset options.

I think both 'combobox' and 'dropdown' are jargon words that we should
not use here.

They also pose a problem for translation: in many of the existing
translations one or both of them were left untranslated (which is
either a mistake, or an even worse case of jargon than in English),
or have been translated into much longer expressions (which worsens
usability of the label).

Change-Id: I9d98acd54a575e05c01040041366afd02567fe75