themes: Unify padded PanelLayout padding
Padded PanelLayouts are foremost used on forms in panels in our context,
hence the variable name `@padding-panel-form`.
Contents are starting with block level elements with higher letter boxes,
where a reduced top distance provides a more harmonious layout.
We already make use of a similar `padding` in core's
`.mw-htmlform-ooui-wrapper.oo-ui-panelLayout-padded` which can be
removed with this patch.
Also adding `@padding-panel-form` variable and equivalent setting in Apex.

Change-Id: Idec187cdfac23eea3a7e7cf7284ec45a2a9245ed