MultilineTextInput: Add 'force' param to adjustSize
In some cases, the widget may be populated while not visible
(e.g. while in an inactive tab) and so the size is not calculated
correctly. Users should be able to call adjustSize when the
widget comes in to view, and a 'force' param is required to
allow the valCache to be ignored.

Change-Id: I07e452765a9306aaf62cdc875db60ea18d5154d3