Element: Avoid crash when getDocument() is called with `window`
When the current document has frames, `window[ 0 ]`, `window[ 1 ]`
etc. refer to the Window objects of each frame. When the frames are
cross-origin, accessing any properties of those objects causes an
exception.

We were incorrectly treating this case as if `window` was a jQuery
object and `window[ 0 ]` a HTMLElement, accessing its properties and
causing an exception.

Change the order of checks so that Window objects are handled earlier.

Bug: T258256
Change-Id: I3d88853fdf9915d2b08063c80ecaf7d92828a5df