Rewrite insufficient label mixin documentation
Just saying, for example, that "the label" is returned can be
confusing when there are so many possibilities what this means:
Is it plain text or something else? What guarantees are given, if
any? Does it include the wrapper element or not?

This patch focuses exclusively on documentation, without touching
any of the existing behavior.

The "do not call this …" note is not correct. The methods can be
called in any order.

Bug: T249167
Bug: T258337
Change-Id: I518966c1c43ac871adeab3fb85efda19471fe27e