Fix label mixin docs related to {string|Function} types
Functions are resolved early, both when setLabel() is called and
on construction time (calls setLabel as well). There is only one
way the this.label property can contain a function: via a
subclass. But even that is resolved on construction time
(assuming the parent constructor is called).

Therefor getLabel() can never return a function.

Therefor setLabelContent() is never called with a function. It
doesn't support this anyway.

This issue is discussed in Ia90150e.

Change-Id: Id73b3cc77c95ce706a3d7dd040eafd8a88057acf