InputWidget: Fix infusion when something removes the 'oo-ui-inputWidget-input' class
When the '.oo-ui-inputWidget-input' element is missing, we ended
up setting config.$input (and later also this.$input and
this.$accessKeyed and a few other properties) to an empty jQuery
collection, like `$( [] )`. This isn't right, and caused code in
AccessKeyedElement to break when accessing them.

Instead it should leave config.$input unset, so that this.$input
and this.$accessKeyed etc. can be generated from scratch.

Bug: T258428
Change-Id: I67aee6840986fecd7ddaf3da6f53c8bc374c22a1