ComboBoxInputWidget: Only show menu on user triggered events
The idea is to open the dropdown menu only when the user interacts
with the element, not every time setValue is called. onEdit is
the event handler that's triggered on keyup and such. It calls
setValue, where the dropdown menu was toggled before. So the
"only" change in this patch is that the menu is toggled a bit
earlier, and not in a separate thread any more (menu.toggle is
now executed outside of the setTimeout in InputWidget.onEdit).

Bug: T261314
Change-Id: I6693fc861dbadc2601e7b62198775deabe67d20c