Fix ComboBoxInputWidget misbehaving when tabbing
A bugfix was made via I6693fc8, moving the menu.toggle() call
from the onInputChange event handler to onEdit. This fixed a
bug but apparently introduced a new one. It looks like the
onEdit event handler is triggered in many, many situations.
The dropdown menu should only open in a few selected
situations.

Bug: T268139
Change-Id: I69b7dfa13d4f270da52595c5659232f2594fea53