Create RequiredElement mixin and use
This sets the 'required' attribute for client side form
validation.

In most cases we don't show the required indicator as
either the widget has no indicator (Radio/Checkbox) or
is using the indicator for something else (FileInput uses
the indicator as a 'clear file' button). These cases can
be addressed later.

Bug: T269474
Change-Id: Id1dbf062ca7cd5f94ea1b0301943e52588f01a19