Fix index error in DropdownInputWidget
The check with empty() is flawed because PHP also considers
the empty string as well as the string "0" empty. I added
test cases for both situations.

I decided to not make this code "clever", i.e. do not attempt
to fix this error at this level. Yea, an empty label is
strange. It's probably a bug in the calling code. Let's not
hide this bug by e.g. skipping this entry, but render it
exactly as requested.

Bug: T275125
Change-Id: I9c98717a6181b6b9f7dd5273387e2eb3417f9be7